requirements.txt
langchain>=0.0.267
langchain-community>=0.0.16
langchain-core>=0.1.3
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
pgvector>=0.2.3
huggingface-hub>=0.19.4
sentence-transformers>=2.2.2
transformers>=4.35.2
streamlit>=1.28.2
python-dotenv>=1.0.0
pydantic>=2.4.2
numpy>=1.24.4
pypdf>=3.17.1
llama-cpp-python>=0.2.19
langchain-google-genai>=0.0.1
